RoadHazeX_Backend/requirements.txt
flask
flask-cors
torch
torchvision
joblib
scikit-learn
pillow
git+https://github.com/openai/CLIP.git
